MpegTools :
  Déverminage divers d'avertissements liés aux nouveaux compilateurs.
  Changements préparatoires à la nouvelle version de TsMerger.

git-svn-id: http://svn.pouchintv.fr/pouchintv/trunk/MpegTools@386 2c5fd688-289b-411d-a95c-3d4e4b52ee9e

+++ TsChecker/TsCheckerDlg.cpp
@@ -350,7 +350,7 @@ void CTsCheckerDialog::PbPaintOvl(HWND hCtl, UINT uMsg)
 
 	TCHAR			szPct[8];
 	static HFONT	hFont	= static_cast<HFONT>(GetStockObject(ANSI_VAR_FONT)); 
-	UINT			nPos	= SendMessage(hCtl, PBM_GETPOS, 0, NULL);
+	UINT			nPos	= static_cast<UINT>(SendMessage(hCtl, PBM_GETPOS, 0, NULL));
 
 	_stprintf_s(szPct, _T("%u %%"), nPos);
 
@@ -363,7 +363,7 @@ void CTsCheckerDialog::PbPaintOvl(HWND hCtl, UINT uMsg)
 	// Select the variable stock font into the specified device context. 
 	if (hOldFont = static_cast<HFONT>(SelectObject(hDC, hFont))) {
 		SetBkMode(hDC, TRANSPARENT);
-		DrawTextEx(hDC, szPct, _tcslen(szPct), &ps.rcPaint, DT_CENTER|DT_SINGLELINE|DT_VCENTER, NULL);
+		DrawTextEx(hDC, szPct, (int)_tcslen(szPct), &ps.rcPaint, DT_CENTER|DT_SINGLELINE|DT_VCENTER, NULL);
 
 		// Restore the original font.        
 		SelectObject(hDC, hOldFont);
